requirements.txt
git+https://github.com/n3ddu8/cnxns.git#egg=cnxns
pyyaml
